Move postgres clause generation to helper
